Trait swim::emulation_core::datapath::Datapath

pub trait Datapath {
    type RegisterData;

    // Required methods
    fn execute_instruction(&mut self) -> DatapathUpdateSignal;
    fn execute_stage(&mut self) -> DatapathUpdateSignal;
    fn set_register_by_str(&mut self, register: &str, data: Self::RegisterData);
    fn set_fp_register_by_str(
        &mut self,
        register: &str,
        data: Self::RegisterData
    );
    fn initialize(
        &mut self,
        initial_pc: usize,
        instructions: Vec<u32>
    ) -> Result<(), String>;
    fn get_memory(&self) -> &Memory;
    fn get_memory_mut(&mut self) -> &mut Memory;
    fn set_memory(&mut self, ptr: u64, data: u32);
    fn is_halted(&self) -> bool;
    fn halt(&mut self);
    fn reset(&mut self);
    fn as_datapath_ref(&self) -> DatapathRef<'_>;
    fn get_syscall_arguments(&self) -> Syscall;

}

A generic datapath.

This has the ability to execute instructions, and to interface with registers and memory. The generic datapath interface additionally specifies a series of data types that correspond to the types of inputs and outputs associated with that datapath.

This design allows developers to create their own datapaths for other architectures than the one created for the sake of this project (MipsDatapath).

Required Associated Types

type RegisterData

The type of data stored within registers. (Suggestions may include u16, u32, or u64 for 16-bit, 32-bit, or 64-bit registers, respectively.)

Required Methods

fn execute_instruction(&mut self) -> DatapathUpdateSignal

Execute a single instruction based on the current state of the datapath. Should the datapath support stages, if the datapath is midway through a stage, the current instruction will be finished instead of executing a new instruction. Should the datapath be in a “halted” state, behavior is undefined.

fn execute_stage(&mut self) -> DatapathUpdateSignal

Execute a single stage of execution based on the current state of the datapath. Should the datapath not support stages, assume the same behavior as Self::execute_instruction(). Should the datapath be in a “halted” state, behavior is undefined.

fn set_register_by_str(&mut self, register: &str, data: Self::RegisterData)

Sets the data in the GP register indicated by the provided string. If it doesn’t exist, this function returns Err.

fn set_fp_register_by_str(&mut self, register: &str, data: Self::RegisterData)

fn initialize( &mut self, initial_pc: usize, instructions: Vec<u32> ) -> Result<(), String>

Reset the datapath, load instructions into memory, and un-sets the is_halted flag. If the process fails, an Err is returned.

fn get_memory(&self) -> &Memory

Retrieve all memory as-is.

fn get_memory_mut(&mut self) -> &mut Memory

Retrieve a mutable reference to memory.

fn set_memory(&mut self, ptr: u64, data: u32)

fn is_halted(&self) -> bool

Returns if the datapath is in a “halted” or “stopped” state. This may be true in the case where an error had occurred previously.

fn halt(&mut self)

Halts the datapath. In order to un-halt the datapath, either call reset() or initialize().

fn reset(&mut self)

Restore the datapath to its default state.

fn as_datapath_ref(&self) -> DatapathRef<'_>

Obtain a reference to the concrete datapath type. Used when datapath-specific logic is needed while dealing with a datapath as a trait object.

fn get_syscall_arguments(&self) -> Syscall

Returns the type of syscall and its arguments (if any exist for that syscall). See the syscall enum for a list of syscalls and what they do.