Enum swim::emulation_core::mips::control_signals::AluControl

pub enum AluControl {

    Addition,
    Subtraction,
    SetOnLessThanSigned,
    SetOnLessThanUnsigned,
    And,
    Or,
    LeftShift16,
    Not,
    MultiplicationSigned,
    MultiplicationUnsigned,
    DivisionSigned,
    DivisionUnsigned,
    AddWithNoWriteOnOverflow,

}

The output of the ALU control unit that directly controls the ALU.

This is not to be confused with the AluOp signal. ALUControl is a function of both AluOp and the funct field of an R-type instruction. While this is a separate control signal, the RegWidth control signal additionally acts as the leading bit of ALUControl. The leading bit of the signal determines the size of the input and output data within the datapath. See RegWidth for more details.

Variants

Addition

_0000 (0) - Perform an addition. (Also used in cases where the ALU result does not matter.) Will not set any overflow signal on overflow.

Subtraction

_0001 (1) - Perform a subtraction. Will not set any underflow signal on underflow.

SetOnLessThanSigned

_0010 (2) - Perform a “set on less than” operation.

SetOnLessThanUnsigned

_0011 (3) - Perform a “set on less than unsigned” operation.

And

_0100 (4) - Perform a bitwise “AND” operation.

Or

_0101 (5) - Perform a bitwise “OR” operation.

LeftShift16

_0110 (6) - Left shift the sign-extended immediate value 16 bits.

Not

_0111 (7) - Perform a bitwise “NOT” operation.

MultiplicationSigned

_1000 (8) - Perform signed multiplication.

MultiplicationUnsigned

_1001 (9) - Perform unsigned multiplication.

DivisionSigned

_1010 (10) - Perform signed integer division. (Returns the integer quotient.)

DivisionUnsigned

_1011 (11) - Perform unsigned integer division. (Returns the integer quotient.)

AddWithNoWriteOnOverflow

_1100 (12) - Perform an addition, and set the OverflowWriteBlock signal in the event of an overflow.