Enum swim::emulation_core::mips::control_signals::floating_point::FpuAluOp

pub enum FpuAluOp {
    AdditionOrEqual,
    Subtraction,
    MultiplicationOrSlt,
    DivisionOrSle,
    And,
    Or,
    Sngt,
    Snge,
}

This doubly determines the operations sent to the floating-point ALU and the floating-point comparator.

Only one of these units are effectively utilized in any given instruction.

The fourth bit of the control signal represents either a single-precision floating-point operation (0), or a double-precision floating-point operation (1). This fourth bit is determined by FpuRegWidth.

Variants

AdditionOrEqual

_000 (0):

ALU: Perform an addition.

Comparator: Set if equal.

Subtraction

_001 (1):

ALU: Perform a subtraction.

MultiplicationOrSlt

_010 (2):

ALU: Perform a multiplication.

Comparator: Set if less than.

DivisionOrSle

_011 (3):

ALU: Perform a division.

Comparator: Set if less than or equal.

And

_100 (4):

ALU: Perform an “AND” operation.

Or

_101 (5):

ALU: Perform an “OR” operation.

Sngt

_110 (6):

Comparator: Set if not greater than.

Snge

_111 (7):

Comparator: Set if not greater than or equal.