Struct swim::emulation_core::mips::coprocessor::MipsFpCoprocessor

pub struct MipsFpCoprocessor {
    instruction: Instruction,
    pub signals: FpuControlSignals,
    pub state: FpuState,
    pub fpr: [u64; 32],
    condition_code: u64,
    data: u64,
}

An implementation of a floating-point coprocessor for the MIPS64 ISA.

Different from the main processor, much of the functionality of the coprocessor is controlled remotely using its available API calls.

Fields
instruction: Instructionsignals: FpuControlSignalsstate: FpuStatefpr: [u64; 32]condition_code: u64data: u64
Implementations

impl MipsFpCoprocessor

pub fn stage_instruction_decode(&mut self)

pub fn stage_execute(&mut self)

pub fn stage_memory(&mut self)

pub fn stage_writeback(&mut self)

pub fn set_instruction(&mut self, instruction_bits: u32)

Set the internally-stored copy of the current instruction. This effectively operates in lieu of any “instruction fetch” functionality since the coprocessor does not fetch instructions.

fn instruction_decode(&mut self)

Decode an instruction into its individual fields.

fn set_reg_width(&mut self)

Set the FpuRegWidth control signal based on the fmt field in the instruction.

pub fn set_data_from_main_processor(&mut self, data: u64)

Sets the data line between the main processor and the Data register. This is then used if deciding data from the main processor should go into the Data register.

pub fn get_data_register(&mut self) -> u64

Gets the contents of the data line between the Data register and the multiplexer in the main processor controlled by the DataWrite control signal.

pub fn set_fp_register_data_from_main_processor(&mut self, data: u64)

Sets the data line between the multiplexer controlled by MemToReg in the main processor and the multiplexer controlled by FpuMemToReg in the floating-point coprocessor.

fn set_control_signals(&mut self)

Set the control signals of the processor based on the instruction opcode and function control signals.

fn read_registers(&mut self)

Read the registers as specified from the instruction and pass the data into the datapath.

fn alu(&mut self)

Perform an ALU operation.

fn comparator(&mut self)

Perform a comparison.

fn write_data(&mut self)

Write to the Data register. This register is used to transfer data between the main processor and the coprocessor.

fn write_condition_code(&mut self)

Set the condition code (CC) register based on the result from the comparator.

fn register_write(&mut self)

Write data to the floating-point register file.