Struct swim::emulation_core::mips::datapath::DatapathState

pub struct DatapathState {

    pub instruction: u32,
    pub rs: u32,
    pub rt: u32,
    pub rd: u32,
    pub shamt: u32,
    pub funct: u32,
    pub imm: u32,
    pub alu_input1: u64,
    pub alu_input2: u64,
    pub alu_result: u64,
    pub data_result: u64,
    pub write_register_destination: usize,
    pub jump_address: u64,
    pub lower_26: u32,
    pub lower_26_shifted_left_by_2: u32,
    pub mem_mux1_to_mem_mux2: u64,
    pub memory_data: u64,
    pub new_pc: u64,
    pub pc_plus_4: u64,
    pub read_data_1: u64,
    pub read_data_2: u64,
    pub register_write_data: u64,
    pub relative_pc_branch: u64,
    pub sign_extend: u64,
    pub sign_extend_shift_left_by_2: u64,
    pub write_data: u64,

}

A collection of all the data lines and wires in the datapath.

Fields
instruction: u32

Data line. The currently loaded instruction. Initialized after the Instruction Fetch stage.

rs: u32rt: u32rd: u32shamt: u32funct: u32imm: u32alu_input1: u64

Data line. The first input of the ALU.

alu_input2: u64

Data line. The second input of the ALU.

alu_result: u64

Data line. The final result as provided by the ALU. Initialized after the Execute stage.

data_result: u64

Data line. The data after the MemToReg multiplexer, but before the DataWrite multiplexer in the main processor.

write_register_destination: usizejump_address: u64

Jump address line. This line carries the concatenation of the high 36 bits of the PC, and lower_26_shifted_left_by_2.

lower_26: u32

Jump 26 bit line. The lower 26 bits of the instruction reserved for the location used by a J-type instruction.

lower_26_shifted_left_by_2: u32

Lower 26 << 2 line. This line carries the lower 28 bits of the jump address.

mem_mux1_to_mem_mux2: u64

Data line. Determines the next value of the PC, given that the current instruction is not a jump.

memory_data: u64

Data line. The data retrieved from memory. Initialized after the Memory stage.

new_pc: u64

New PC line. In the WB stage, this line is written to the PC.

pc_plus_4: u64

Data line. Contains PC + 4.

read_data_1: u64

Data line. Data read from the register file based on the rs field of the instruction. Initialized after the Instruction Decode stage.

read_data_2: u64

Data line. Data read from the register file based on the rt field of the instruction. Initialized after the Instruction Decode stage.

register_write_data: u64

Data line. The data after the DataWrite multiplexer in the main processor and the main processor register file.

relative_pc_branch: u64

Data line. New PC value used if branching is set for an instruction.

sign_extend: u64

Data line. The instruction’s immediate value sign-extended to 64 bits. Initialized after the Instruction Decode stage.

sign_extend_shift_left_by_2: u64

Data line. The sign_extend line, shifted left by two bits.

write_data: u64

Data line. The data that will be written to memory.