Struct swim::emulation_core::mips::datapath::MipsDatapath

pub struct MipsDatapath {
    pub registers: GpRegisters,
    pub memory: Memory,
    pub coprocessor: MipsFpCoprocessor,
    pub instruction: Instruction,
    pub signals: ControlSignals,
    pub datapath_signals: DatapathSignals,
    pub state: DatapathState,
    pub current_stage: Stage,
    is_halted: bool,
}

An implementation of a datapath for the MIPS64 ISA.

Fields
registers: GpRegistersmemory: Memorycoprocessor: MipsFpCoprocessorinstruction: Instructionsignals: ControlSignalsdatapath_signals: DatapathSignalsstate: DatapathStatecurrent_stage: Stage

The currently-active stage in the datapath.

is_halted: bool

Boolean value that states whether the datapath has halted.

This is set in the event of any syscall instruction. To unset this, Self::initialize() should be used.

Implementations

impl MipsDatapath

pub fn initialize(&mut self, instructions: Vec<u32>) -> Result<(), String>

Reset the datapath, load instructions into memory, and un-sets the is_halted flag. If the process fails, an Err is returned.

fn load_instructions(&mut self, instructions: Vec<u32>) -> Result<(), String>

Load a vector of 32-bit instructions into memory. If the process fails, from a lack of space or otherwise, an Err is returned.

pub fn error(&mut self, _message: &str)

Handle an otherwise irrecoverable error within the datapath.

fn stage_instruction_fetch(&mut self)

Stage 1 of 5: Instruction Fetch (IF)

Fetch the current instruction based on the given PC and load it into the datapath.

fn stage_instruction_decode(&mut self)

Stage 2 of 5: Instruction Decode (ID)

Parse the instruction, set control signals, and read registers.

If the instruction is determined to be a syscall, immediately finish the instruction and set the is_halted flag.

fn stage_execute(&mut self)

Stage 3 of 5: Execute (EX)

Execute the current instruction with some arithmetic operation.

fn stage_memory(&mut self)

Stage 4 of 5: Memory (MEM)

Read or write to memory.

fn stage_writeback(&mut self)

Stage 5 of 5: Writeback (WB)

Write the result of the instruction’s operation to a register, if desired. Additionally, set the PC for the next instruction.

fn instruction_fetch(&mut self)

Load the raw binary instruction from memory and into the datapath. If there is an error with loading the word, assume the instruction to be bitwise zero and error.

fn pc_plus_4(&mut self)

fn instruction_decode(&mut self)

Decode an instruction into its individual fields.

fn sign_extend(&mut self)

Extend the sign of a 16-bit value to the other 48 bits of a 64-bit value.

fn set_control_signals(&mut self)

Set the control signals for the datapath based on the instruction’s opcode.

fn set_rtype_control_signals(&mut self, r: RType)

Set rtype control signals. This function may have a Match statement added in the future for dealing with different special case rtype instructions.

fn set_itype_control_signals(&mut self, i: IType)

Set the control signals for the datapath, specifically in the case where the instruction is an I-type.

fn set_jtype_control_signals(&mut self, j: JType)

Set control signals for J-Type instructions

fn set_fpu_reg_imm_control_signals(&mut self, i: FpuRegImmType)

Set the control signals for the datapath, specifically in the case where the instruction is an FPU register-immediate type.

fn set_fpu_itype_control_signals(&mut self, i: FpuIType)

Set the control signals for the datapath, specifically in the case where the instruction is an FPU I-type.

fn read_registers(&mut self)

Read the registers as specified from the instruction and pass the data into the datapath.

fn set_alu_control(&mut self)

Set the ALU control signal based on the AluOp signal.

fn shift_lower_26_left_by_2(&mut self)

fn construct_jump_address(&mut self)

fn alu(&mut self)

Perform an ALU operation.

Implementation Note: Unlike the MIPS64 specification, this ALU does not handle exceptions due to integer overflow. However, it will set the OverflowWriteBlock signal on overflow when the operation AluControl is set to AddWithNoWriteOverflow.

fn calc_relative_pc_branch(&mut self)

fn calc_cpu_branch_signal(&mut self)

Determine the value of the CpuBranch signal.

fn memory_read(&mut self)

Read from memory based on the address provided by the ALU in DatapathState::alu_result. Returns the result to DatapathState::memory_data. Should the address be invalid or otherwise memory cannot be read at the given address, bitwise 0 will be used in lieu of any data.

fn memory_write(&mut self)

Write to memory based on the address provided by the ALU in DatapathState::alu_result. The source of the data being written to memory is determined by MemWriteSrc.

fn calc_general_branch_signal(&mut self)

fn pick_pc_plus_4_or_relative_branch_addr_mux1(&mut self)

fn set_new_pc_mux2(&mut self)

fn register_write(&mut self)

Write to a register. This will only write if the RegWrite control signal is set.

fn set_pc(&mut self)

Update the program counter register.

This function is called from the WB stage.