Module swim::emulation_core::mips::registers

Register structure and API.

Structs

GpRegisterTypeIter

An iterator over the variants of Self

GpRegisters

Collection of general-purpose registers used by the datapath.

GpRegistersIter

Iterator that is used to view each register in the register file.

Enums

GpRegisterType

Specifies all of the valid registers accessible in an instance of GpRegisters.