Module swim::emulation_core::riscv::control_signals

Internal datapath control signals.

Modules

floating_point

Floating Point Control Signals Module.

Structs

ControlSignals

Full collection of control signals.

Enums

AluOp

The output of the ALU control unit that directly controls the ALU.

BranchJump

Selection of Branch Operations.

ImmSelect

Selection of different Immediate forms.

MemWriteSrc

Determines, given that a write value in ReadWrite is set, the source of the data will be written to memory.

OP1Select

Selection of different sources for the first operand.

OP2Select

Selection of different sources for the second operand.

ReadWrite

Selection of Read and Write Operations.

RegDst

Determines, given that RegWriteEn is set, which destination register to write to, which largely depends on the instruction format.

RegWriteEn

Determines if the register file should be written to.

SysOp

Selection of System Operations.

WBSel

Determines, given RegWriteEn is set, what the source of a register’s new data will be.