Module swim::emulation_core::riscv::control_signals::floating_point

Floating Point Control Signals Module.

Structs

FpuControlSignals

Enums

DataSrc

Determines the source of the Data register in the floating-point unit.

DataWrite

Determines whether to write to the Data register in the floating-point unit.

FpuAluOp

This doubly determines the operations sent to the floating-point ALU and the floating-point comparator.

FpuMemToReg

Determines, given that FpuRegWrite is set, what the source of a floating-point register’s new data will be.

FpuRegDst

Determines, given that FpuRegWrite is set, which destination register to write to, which largely depends on the instruction format.

FpuRegWrite

Determines if the floating-point register file should be written to.

RoundingMode

Selection of Rounding Modes