Module swim::emulation_core::riscv::datapath

Implementation of a RISC-V datapath.

It is assumed that while moving through stages, only one instruction will be active any any given point in time. Due to this, we consider the datapath to be a “pseudo-single-cycle datapath.”

For the most part, this datapath is an implementation of RISC-V Spec 2.2 with extensions: RV32I, RV64I, M, and F.

Notes on is_halted

The datapath starts with the is_halted flag set.

RiscDatapath::initialize() should be used to un-set is_halted.

The EBREAK instruction simply performs a no-operation instruction, except for setting the boolean flag is_halted.

Invalid instructions will cause the datapath to set the is_halted flag.

Structs

RiscDatapath

An implementation of a datapath for the MIPS64 ISA.

RiscDatapathState

A collection of all the data lines and wires in the datapath.

Enums

RiscStage

The possible stages the datapath could be in during execution.