Function swim::parser::operand_reading::read_immediate

pub fn read_immediate(
    given_text: &str,
    operand_number: i32,
    num_bits: u32
) -> (u32, Option<Error>)

This function takes a string representation of an immediate value and the number of bits available to represent it and attempts to translate it to an actual integer. If the value cannot be cast to int or is too big to be represented by the available bits, an error is returned.