Module swim::parser::parser_structs_and_enums

Structs

Data

A collection of all relevant information found about a variable in the Parser/Assembler

Error

FPRegister

FPRegisterRiscv

GPRegister

GPRegisterRiscv

Instruction

A collection of all relevant information found about an instruction in the Parser/Assembler

InstructionDescription

LabelInstance

MonacoLineInfo

This struct holds all the information we gather in the parser & assembler about a single line the user wrote

ProgramInfo

Wrapper for all information gathered in the Parser/Assembler about the written program.

PseudoDescription

Token

Enums

ErrorType

OperandType

RegisterType

TokenType

Constants

FP_REGISTERS

Contains every MIPS floating point register name and binary value. Any reference to fp registers throughout the parser/assembler should reference this array

GP_REGISTERS

Contains every MIPS general purpose register’s binary value and the various names they are recognized as. Any reference to gp registers throughout the parser/assembler should reference this array

RISCV_FP_REGISTERS

Contains every RISC-V floating point register name and binary value. Any reference to fp registers throughout the parser/assembler should reference this array

RISCV_GP_REGISTERS

Contains every RISC-V general purpose register’s binary value and the various names they are recognized as. Any reference to gp registers throughout the parser/assembler should reference this array

SUPPORTED_INSTRUCTIONS_MIPS

SUPPORTED_INSTRUCTIONS_RISCV

UNSUPPORTED_INSTRUCTIONS_MIPS

UNSUPPORTED_INSTRUCTIONS_RISCV

Functions

print_data_contents

print_instruction_contents

print_vec_of_data

print_vec_of_instructions