Module swim::parser::parsing

Functions

create_label_map

Create_label_map builds a hashmap of addresses for labels in memory

remove_commas

Checks the name of every token on a line and makes sure that labels, directives, and operators do not end in commas while all but the last operand or datum on a line does. Also, pops commas off of the end of all tokens on the line.

separate_data_and_text

This function takes the vector of lines created by tokenize program and turns them into instructions assigning labels, operators, operands, and line numbers and data assigning labels, data types, and values

suggest_error_corrections

Goes through each error found in the parsing & assembling process and suggests to the user a way of correcting the error. This error message is attached to the corresponding instruction or data and monaco line info, and compiled into a string to be output to the console and returns that

tokenize_program

Takes the initial string of the program given by the editor and turns it into a vector of Line, a struct that holds tokens and the original line number.