Module swim::ui::visual_datapath

The visual datapath component.

This component operates largely in JS-land, manipulating the page’s DOM directly. The visual datapath operates by instantiating an <object> element that contains the SVG diagram of the datapath. This offers direct access to its own sub-DOM that can be manipulated, including but not limited to changing the colors of individual lines and components in the datapath.

At its core, interactivity is enabled through event listeners. Within the VisualDatapath struct is a shared reference to a Vec<EventListener>, which contains the currently active event listeners at any given point. Whenever a set of lines are highlighted in the datapath, event listeners are created for each of those lines as well. These event listeners wait for mouse movement over lines to determine when, where, and what information should be displayed.

Any existing event listeners need to be disposed of when the diagram updates and a new set of lines are highlighted. This is handled during the changed() function built into the Yew framework. After this, it will proceed to re-render the diagram and create new event listeners.

Displaying information about a line is done via a “popup” element, which is moved around the page depending on the position of the mouse. Its visibility is controlled by the aforementioned event listeners.

Modules

consts

utils

Helpful common functions used for the visual datapath.

Structs

VisualDatapath

VisualDatapathProps

Enums

DatapathSize