Struct swim::ui::visual_datapath::VisualDatapath

pub struct VisualDatapath {
    active_listeners: Rc<RefCell<Vec<EventListener>>>,
}

Fields
active_listeners: Rc<RefCell<Vec<EventListener>>>
Implementations

impl VisualDatapath

pub fn add_event_listeners(&mut self, new_listeners: Vec<EventListener>)

Move the contents of new_listeners to self.event_listeners.

pub fn clear_event_listeners(&mut self)

Clear the contents of self.event_listeners.

pub fn initialize(&mut self)

Make the SVG element on the page interact-able.

There are aspects about the SVG element from exporting that initially make it hard to properly interact with. This makes a few adjustments to the element to allow this functionality.

This function is written in the way that it is (encapsulating the initialization within a Callback) due to the nature of page loading.

While the <VisualDatapath> element is finished loading on the page when the rendered() function is called, there is a likely chance that the virtual DOM within the <object> element has not yet finished loading. This is circumvented by creating an event listener on the <object> element for the “load” event, which will guarantee when that virtual DOM is actually ready to be manipulated.

pub fn highlight_stage(
datapath: MipsDatapath,
nodes: &HtmlCollection,
stage: String
) -> Result<Vec<EventListener>, JsValue>

Highlight and enable interactivity for all lines in a given stage.

Consequently un-highlights for all other lines. (Disabling interactivity for other lines is done within Self::changed().)

Takes as input a reference to an HtmlCollection that contains all the <g> elements of the SVG diagram.

Valid strings in the stage parameter:

instruction_fetch

instruction_decode

execute

memory

writeback

If successful, returns a Vec<EventListener> containing the new event listeners generated.

pub fn activate_element(
datapath: MipsDatapath,
element: &Element
) -> Result<Vec<EventListener>, JsValue>

Activates a given <g> element.

If successful, returns a Vec<EventListener> containing the newly created event listeners.

pub fn deactivate_element(element: &Element) -> Result<(), JsValue>

Deactivates a given <g> element.

pub fn set_inactive(path: &HtmlElement) -> Result<(), JsValue>

Set a line to be inactive.

pub fn set_active_unhovered(path: &HtmlElement) -> Result<(), JsValue>

Set a line to be active and unhovered.

pub fn set_active_hovered(path: &HtmlElement) -> Result<(), JsValue>

Set a line to be active and hovered.

pub fn set_color(path: &HtmlElement, color: &str) -> Result<(), JsValue>

Set a line’s color.