Module swim::ui::visual_datapath::utils

Helpful common functions used for the visual datapath.

Functions

calculate_popup_position

Given the mouse location, the size of the popup, and the window size, return the coordinates of the top left corner of where the popup should go.

do_over_html_collection

Perform some function over an HtmlCollection, assuming each element inside of it is valid.

do_over_html_collection_safe

Perform some function over an HtmlCollection, but without unwrapping each element.

get_datapath_iframe_mouse_position

Given a MouseEvent inside the datapath <object> element, returns the coordinates of the mouse on the full screen.

get_datapath_iframe_size

get_datapath_position

Returns the relative coordinates of the <object> element to the page.

get_datapath_root

Returns an HtmlObjectElement corresponding to the <object> element in HTML.

get_g_elements

Returns an HtmlCollection containing all the <g> elements within the SVG diagram.

get_popup_element

Returns an HtmlElement corresponding to the <div id="popup"> element in HTML.

get_window_size

Returns the size of the browser window in pixels.

populate_popup_information

Set the data contained in the popup.

u64_to_bits

Convert an integer value to a string, limited to bits number of bits.

visual_line_to_data