Module swim::emulation_core::mips::control_signals::floating_point

Structs

FpuControlSignals

Enums

Cc

Determines, given that CcWrite is set, which condition code register should be written to or read from for a given operation.

CcWrite

Determines if the condition code register file should be written to.

DataSrc

Determines the source of the Data register in the floating-point unit.

DataWrite

Determines whether to write to the Data register in the floating-point unit.

FpuAluOp

This doubly determines the operations sent to the floating-point ALU and the floating-point comparator.

FpuBranch

Determines if the floating-point unit should consider branching, based on the contents of the condition code register.

FpuMemToReg

Determines, given that FpuRegWrite is set, what the source of a floating-point register’s new data will be.

FpuRegDst

Determines, given that FpuRegWrite is set, which destination register to write to, which largely depends on the instruction format.

FpuRegWidth

Determines the amount of data to be sent or received from registers and the ALU.

FpuRegWrite

Determines if the floating-point register file should be written to.