Module swim::emulation_core::mips::datapath

Implementation of a MIPS64 datapath.

It is assumed that while moving through stages, only one instruction will be active any any given point in time. Due to this, we consider the datapath to be a “pseudo-single-cycle datapath.”

For the most part, this datapath is an implementation of MIPS64 Version 6. (See below for exceptions.)

Differences Compared to MIPS64 Version 6

It should be noted that this datapath chooses to diverge from the MIPS64 version 6 specification for the sake of simplicity in a few places:

There is no exception handling, including that for integer overflow. (See MipsDatapath::alu() and the following bullet.)

The add, addi, dadd, daddi, sub, and dsub instructions do not follow the proper MIPS specification in terms of integer overflow/wraparound. That is, if there is integer wraparound, the general-purpose register should not be written to. In our implementation, the general-purpose register is written to regardless.

32-bit instructions are treated exclusively with 32 bits, and the upper 32 bits stored in a register are completely ignored in any of these cases. For example, before an add instruction, it should be checked whether it is a sign-extended 32-bit value stored in a 64-bit register. Instead, the upper 32 bits are ignored when being used for 32-bit instructions.

Instead of implementing the cmp.condn.fmt instructions, this datapath implements the c.cond.fmt instructions from MIPS64 version 5.

Unlike MIPS specification, SWIM only uses 1 condition code register (cc), rather than offering 8 condition code registers. The datapath will assume that the cc field in a floating-point comparison or floating-point branch instruction is 0.

This datapath implements the addi instruction as it exists in MIPS64 version 5. This instruction was deprecated in MIPS64 version 6 to allow for the beqzalc, bnezalc, beqc, and bovc instructions.

This datapath implements daddi as it exists in MIPS64 version 5. This instruction was deprecated in MIPS64 version 6.

Unlike the MIPS64 version 6 specification for the jal and jalr instructions, PC + 4 is stored in GPR[31], not PC + 8, as there is no implementation of branch delay slots.

Notes on is_halted

The datapath starts with the is_halted flag set.

MipsDatapath::initialize() should be used to un-set is_halted.

The syscall instruction simply performs a no-operation instruction, except for setting the boolean flag is_halted.

Invalid instructions will cause the datapath to set the is_halted flag.

Structs

DatapathState

A collection of all the data lines and wires in the datapath.

MipsDatapath

An implementation of a datapath for the MIPS64 ISA.

Enums

Stage

The possible stages the datapath could be in during execution.