Struct swim::emulation_core::mips::datapath::DatapathState

pub struct DatapathState {

    instruction: u32,
    rs: u32,
    rt: u32,
    rd: u32,
    shamt: u32,
    funct: u32,
    imm: u32,
    read_data_1: u64,
    read_data_2: u64,
    sign_extend: u64,
    alu_result: u64,
    memory_data: u64,
    data_result: u64,
    register_write_data: u64,
    lower_26: u32,
    lower_26_shifted_left_by_2: u32,
    jump_address: u64,
    pc_plus_4: u64,
    new_pc: u64,
    relative_pc_branch: u64,
    mem_mux1_to_mem_mux2: u64,

}

A collection of all the data lines and wires in the datapath.

Fields
instruction: u32

Data line. The currently loaded instruction. Initialized after the Instruction Fetch stage.

rs: u32rt: u32rd: u32shamt: u32funct: u32imm: u32read_data_1: u64

Data line. Data read from the register file based on the rs field of the instruction. Initialized after the Instruction Decode stage.

read_data_2: u64

Data line. Data read from the register file based on the rt field of the instruction. Initialized after the Instruction Decode stage.

sign_extend: u64

Data line. The instruction’s immediate value sign-extended to 64 bits. Initialized after the Instruction Decode stage.

alu_result: u64

Data line. The final result as provided by the ALU. Initialized after the Execute stage.

memory_data: u64

Data line. The data retrieved from memory. Initialized after the Memory stage.

data_result: u64

Data line. The data after the MemToReg multiplexer, but before the DataWrite multiplexer in the main processor.

register_write_data: u64

Data line. The data after the DataWrite multiplexer in the main processor and the main processor register file.

lower_26: u32

Jump 26 bit line. The low 26 bits of the instruction reserved for possible use by the J instruction

lower_26_shifted_left_by_2: u32

Lower 26 << 2 line. This line carries the low 28 bits for the jump address

jump_address: u64

Jump address line. The line the will carry the combination of the high 4 bits and pc, and the lower_26_for_jump_line bits shifted left by 2.

pc_plus_4: u64

PC + 4 line. Yeah, just PC + 4

new_pc: u64

New PC line. In the WB stage this line is written to registers.pc

relative_pc_branch: u64

*Relative PC branch address line

mem_mux1_to_mem_mux2: u64

bla bla bal